+++ src/PhpSpreadsheet/Cell/CellAddress.php
@@ -27,7 +27,7 @@ class CellAddress implements Stringable
 
     public function __destruct()
     {
-        $this->worksheet = null;
+        unset($this->worksheet);
     }
 
     /**
@@ -45,21 +45,19 @@ class CellAddress implements Stringable
     {
         self::validateColumnAndRow($columnId, $rowId);
 
-        /** @phpstan-ignore-next-line */
-        return new static(Coordinate::stringFromColumnIndex($columnId) . ((string) $rowId), $worksheet);
+        return new self(Coordinate::stringFromColumnIndex($columnId) . ((string) $rowId), $worksheet);
     }
 
     public static function fromColumnRowArray(array $array, ?Worksheet $worksheet = null): self
     {
         [$columnId, $rowId] = $array;
 
-        return static::fromColumnAndRow($columnId, $rowId, $worksheet);
+        return self::fromColumnAndRow($columnId, $rowId, $worksheet);
     }
 
     public static function fromCellAddress(mixed $cellAddress, ?Worksheet $worksheet = null): self
     {
-        /** @phpstan-ignore-next-line */
-        return new static($cellAddress, $worksheet);
+        return new self($cellAddress, $worksheet);
     }
 
     /**
@@ -115,7 +113,7 @@ class CellAddress implements Stringable
             $newRowId = 1;
         }
 
-        return static::fromColumnAndRow($this->columnId, $newRowId);
+        return self::fromColumnAndRow($this->columnId, $newRowId);
     }
 
     public function previousRow(int $offset = 1): self
@@ -130,7 +128,7 @@ class CellAddress implements Stringable
             $newColumnId = 1;
         }
 
-        return static::fromColumnAndRow($newColumnId, $this->rowId);
+        return self::fromColumnAndRow($newColumnId, $this->rowId);
     }
 
     public function previousColumn(int $offset = 1): self