+++ src/PhpSpreadsheet/Cell/StringValueBinder.php
@@ -3,8 +3,10 @@
 namespace PhpOffice\PhpSpreadsheet\Cell;
 
 use DateTimeInterface;
+use PhpOffice\PhpSpreadsheet\Exception as SpreadsheetException;
 use PhpOffice\PhpSpreadsheet\RichText\RichText;
 use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use Stringable;
 
 class StringValueBinder implements IValueBinder
 {
@@ -82,6 +84,9 @@ class StringValueBinder implements IValueBinder
         if (is_object($value)) {
             return $this->bindObjectValue($cell, $value);
         }
+        if ($value !== null && !is_scalar($value)) {
+            throw new SpreadsheetException('Unable to bind unstringable ' . gettype($value));
+        }
 
         // sanitize UTF-8 strings
         if (is_string($value)) {
@@ -111,13 +116,14 @@ class StringValueBinder implements IValueBinder
         // Handle any objects that might be injected
         if ($value instanceof DateTimeInterface) {
             $value = $value->format('Y-m-d H:i:s');
+            $cell->setValueExplicit($value, DataType::TYPE_STRING);
         } elseif ($value instanceof RichText) {
             $cell->setValueExplicit($value, DataType::TYPE_INLINE);
-
-            return true;
+        } elseif ($value instanceof Stringable) {
+            $cell->setValueExplicit((string) $value, DataType::TYPE_STRING);
+        } else {
+            throw new SpreadsheetException('Unable to bind unstringable object of type ' . get_class($value));
         }
-
-        $cell->setValueExplicit((string) $value, DataType::TYPE_STRING); // @phpstan-ignore-line
 
         return true;
     }