+++ src/PhpSpreadsheet/Helper/Sample.php
@@ -198,7 +198,7 @@ class Sample
      *
      * @codeCoverageIgnore
      */
-    public function renderChart(Chart $chart, string $fileName): void
+    public function renderChart(Chart $chart, string $fileName, ?Spreadsheet $spreadsheet = null): void
     {
         if ($this->isCli() === true) {
             return;
@@ -206,17 +206,32 @@ class Sample
         Settings::setChartRenderer(MtJpGraphRenderer::class);
 
         $fileName = $this->getFilename($fileName, 'png');
+        $title = $chart->getTitle();
+        $caption = null;
+        if ($title !== null) {
+            $calculatedTitle = $title->getCalculatedTitle($spreadsheet);
+            if ($calculatedTitle !== null) {
+                $caption = $title->getCaption();
+                $title->setCaption($calculatedTitle);
+            }
+        }
 
         try {
             $chart->render($fileName);
             $this->log('Rendered image: ' . $fileName);
-            $imageData = file_get_contents($fileName);
+            $imageData = @file_get_contents($fileName);
             if ($imageData !== false) {
                 echo '<div><img src="data:image/gif;base64,' . base64_encode($imageData) . '" /></div>';
+            } else {
+                $this->log('Unable to open chart' . PHP_EOL);
             }
         } catch (Throwable $e) {
             $this->log('Error rendering chart: ' . $e->getMessage() . PHP_EOL);
         }
+        if (isset($title, $caption)) {
+            $title->setCaption($caption);
+        }
+        Settings::unsetChartRenderer();
     }
 
     public function titles(string $category, string $functionName, ?string $description = null): void