+++ src/PhpSpreadsheet/Reader/Xlsx/ConditionalStyles.php
@@ -3,7 +3,9 @@
 namespace PhpOffice\PhpSpreadsheet\Reader\Xlsx;
 
 use PhpOffice\PhpSpreadsheet\Reader\Xlsx\Styles as StyleReader;
+use PhpOffice\PhpSpreadsheet\Style\Color;
 use PhpOffice\PhpSpreadsheet\Style\Conditional;
+use PhpOffice\PhpSpreadsheet\Style\ConditionalFormatting\ConditionalColorScale;
 use PhpOffice\PhpSpreadsheet\Style\ConditionalFormatting\ConditionalDataBar;
 use PhpOffice\PhpSpreadsheet\Style\ConditionalFormatting\ConditionalFormattingRuleExtension;
 use PhpOffice\PhpSpreadsheet\Style\ConditionalFormatting\ConditionalFormatValueObject;
@@ -25,11 +27,14 @@ class ConditionalStyles
 
     private array $dxfs;
 
-    public function __construct(Worksheet $workSheet, SimpleXMLElement $worksheetXml, array $dxfs = [])
+    private StyleReader $styleReader;
+
+    public function __construct(Worksheet $workSheet, SimpleXMLElement $worksheetXml, array $dxfs, StyleReader $styleReader)
     {
         $this->worksheet = $workSheet;
         $this->worksheetXml = $worksheetXml;
         $this->dxfs = $dxfs;
+        $this->styleReader = $styleReader;
     }
 
     public function load(): void
@@ -45,13 +50,13 @@ class ConditionalStyles
         $this->worksheet->setSelectedCells($selectedCells);
     }
 
-    public function loadFromExt(StyleReader $styleReader): void
+    public function loadFromExt(): void
     {
         $selectedCells = $this->worksheet->getSelectedCells();
 
         $this->ns = $this->worksheetXml->getNamespaces(true);
         $this->setConditionalsFromExt(
-            $this->readConditionalsFromExt($this->worksheetXml->extLst, $styleReader)
+            $this->readConditionalsFromExt($this->worksheetXml->extLst)
         );
 
         $this->worksheet->setSelectedCells($selectedCells);
@@ -68,7 +73,7 @@ class ConditionalStyles
         }
     }
 
-    private function readConditionalsFromExt(SimpleXMLElement $extLst, StyleReader $styleReader): array
+    private function readConditionalsFromExt(SimpleXMLElement $extLst): array
     {
         $conditionals = [];
         if (!isset($extLst->ext)) {
@@ -110,7 +115,7 @@ class ConditionalStyles
                 $priority = (int) $attributes->priority;
 
                 $conditional = $this->readConditionalRuleFromExt($extCfRuleXml, $attributes);
-                $cfStyle = $this->readStyleFromExt($extCfRuleXml, $styleReader);
+                $cfStyle = $this->readStyleFromExt($extCfRuleXml);
                 $conditional->setStyle($cfStyle);
                 $conditionals[$sqref][$priority] = $conditional;
             }
@@ -146,17 +151,17 @@ class ConditionalStyles
         return $conditional;
     }
 
-    private function readStyleFromExt(SimpleXMLElement $extCfRuleXml, StyleReader $styleReader): Style
+    private function readStyleFromExt(SimpleXMLElement $extCfRuleXml): Style
     {
         $cfStyle = new Style(false, true);
         if ($extCfRuleXml->dxf) {
             $styleXML = $extCfRuleXml->dxf->children();
 
             if ($styleXML->borders) {
-                $styleReader->readBorderStyle($cfStyle->getBorders(), $styleXML->borders);
+                $this->styleReader->readBorderStyle($cfStyle->getBorders(), $styleXML->borders);
             }
             if ($styleXML->fill) {
-                $styleReader->readFillStyle($cfStyle->getFill(), $styleXML->fill);
+                $this->styleReader->readFillStyle($cfStyle->getFill(), $styleXML->fill);
             }
         }
 
@@ -198,6 +203,7 @@ class ConditionalStyles
         $conditionalFormattingRuleExtensions = ConditionalFormattingRuleExtension::parseExtLstXml($extLst);
         $conditionalStyles = [];
 
+        /** @var SimpleXMLElement $cfRule */
         foreach ($cfRules as $cfRule) {
             $objConditional = new Conditional();
             $objConditional->setConditionType((string) $cfRule['type']);
@@ -231,7 +237,11 @@ class ConditionalStyles
 
             if (isset($cfRule->dataBar)) {
                 $objConditional->setDataBar(
-                    $this->readDataBarOfConditionalRule($cfRule, $conditionalFormattingRuleExtensions) // @phpstan-ignore-line
+                    $this->readDataBarOfConditionalRule($cfRule, $conditionalFormattingRuleExtensions)
+                );
+            } elseif (isset($cfRule->colorScale)) {
+                $objConditional->setColorScale(
+                    $this->readColorScale($cfRule)
                 );
             } elseif (isset($cfRule['dxfId'])) {
                 $objConditional->setStyle(clone $this->dxfs[(int) ($cfRule['dxfId'])]);
@@ -270,12 +280,46 @@ class ConditionalStyles
 
         //color
         if (isset($cfRule->dataBar->color)) {
-            $dataBar->setColor((string) $cfRule->dataBar->color['rgb']);
+            $dataBar->setColor($this->styleReader->readColor($cfRule->dataBar->color));
         }
         //extLst
         $this->readDataBarExtLstOfConditionalRule($dataBar, $cfRule, $conditionalFormattingRuleExtensions);
 
         return $dataBar;
+    }
+
+    private function readColorScale(simpleXMLElement|stdClass $cfRule): ConditionalColorScale
+    {
+        $colorScale = new ConditionalColorScale();
+        $types = [];
+        foreach ($cfRule->colorScale->cfvo as $cfvoXml) {
+            $attr = $cfvoXml->attributes() ?? [];
+            $type = (string) ($attr['type'] ?? '');
+            $types[] = $type;
+            $val = $attr['val'] ?? null;
+            if ($type === 'min') {
+                $colorScale->setMinimumConditionalFormatValueObject(new ConditionalFormatValueObject($type, $val));
+            } elseif ($type === 'percentile') {
+                $colorScale->setMidpointConditionalFormatValueObject(new ConditionalFormatValueObject($type, $val));
+            } elseif ($type === 'max') {
+                $colorScale->setMaximumConditionalFormatValueObject(new ConditionalFormatValueObject($type, $val));
+            }
+        }
+        $idx = 0;
+        foreach ($cfRule->colorScale->color as $color) {
+            $type = $types[$idx];
+            $rgb = $this->styleReader->readColor($color);
+            if ($type === 'min') {
+                $colorScale->setMinimumColor(new Color($rgb));
+            } elseif ($type === 'percentile') {
+                $colorScale->setMidpointColor(new Color($rgb));
+            } elseif ($type === 'max') {
+                $colorScale->setMaximumColor(new Color($rgb));
+            }
+            ++$idx;
+        }
+
+        return $colorScale;
     }
 
     /**