+++ tests/PhpSpreadsheetTests/Calculation/StructuredReferenceFormulaTest.php
@@ -41,7 +41,7 @@ class StructuredReferenceFormulaTest extends TestCase
         $result = $spreadsheet->getActiveSheet()->getCell($cellAddress)->getCalculatedValue();
         self::assertSame('Region', $result);
 
-        $spreadsheet->getCalculationEngine()->flushInstance(); // @phpstan-ignore-line
+        $spreadsheet->getCalculationEngine()?->flushInstance();
         $table->setShowHeaderRow(false);
 
         $result = $spreadsheet->getActiveSheet()->getCell($cellAddress)->getCalculatedValue();